add token verification method into inputstreambinder class (otherwise client app needs dependency on encryptionutils class)
