+++ src/main/java/com/nextcloud/android/sso/aidl/NextcloudRequest.java
@@ -19,8 +19,6 @@
 
 package com.nextcloud.android.sso.aidl;
 
-import com.owncloud.android.utils.EncryptionUtils;
-
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.List;
@@ -145,28 +143,5 @@ public class NextcloudRequest implements Serializable {
 
     public boolean isFollowRedirects() {
         return this.followRedirects;
-    }
-
-    public boolean validateToken(String token) {
-        String salt = this.token.split("\\$")[1]; // TODO extract "$"
-
-        String newHash = EncryptionUtils.generateSHA512(token, salt);
-
-        // As discussed with Lukas R. at the Nextcloud Conf 2018, always compare whole strings
-        // and don't exit prematurely if the string does not match anymore to prevent timing-attacks
-        return isEqual(this.token.getBytes(), newHash.getBytes());
-    }
-
-    // Taken from http://codahale.com/a-lesson-in-timing-attacks/
-    private static boolean isEqual(byte[] a, byte[] b) {
-        if (a.length != b.length) {
-            return false;
-        }
-
-        int result = 0;
-        for (int i = 0; i < a.length; i++) {
-            result |= a[i] ^ b[i];
-        }
-        return result == 0;
     }
 }
